lisa.trace: Move Trace.get_peripheral_clock_effective_rate() into frequency analysis
Move domain-specific functions into the corresponding analysis, where it
is used.
