+++ lisa/trace.py
@@ -843,46 +843,6 @@ class Trace(Loggable, TraceBase):
 
             self._ftrace.cpu_frequency.data_frame = df
 
-###############################################################################
-# Utility Methods
-###############################################################################
-
-
-    @memoized
-    def get_peripheral_clock_effective_rate(self, clk_name):
-        logger = self.get_logger()
-        if clk_name is None:
-            logger.warning('no specified clk_name in computing peripheral clock, returning None')
-            return
-        if not self.has_events('clock_set_rate'):
-            logger.warning('Events [clock_set_rate] not found, returning None!')
-            return
-        rate_df = self.df_events('clock_set_rate')
-        enable_df = self.df_events('clock_enable')
-        disable_df = self.df_events('clock_disable')
-        pd.set_option('display.expand_frame_repr', False)
-
-        freq = rate_df[rate_df.clk_name == clk_name]
-        if not enable_df.empty:
-            enables = enable_df[enable_df.clk_name == clk_name]
-        if not disable_df.empty:
-            disables = disable_df[disable_df.clk_name == clk_name]
-
-        freq = pd.concat([freq, enables, disables], sort=False).sort_index()
-        if freq.empty:
-            logger.warning('No events for clock ' + clk_name + ' found in trace')
-            return
-
-        freq['start'] = freq.index
-        freq['len'] = (freq.start - freq.start.shift()).fillna(0).shift(-1)
-        # The last value will be NaN, fix to be appropriate length
-        freq.loc[freq.index[-1], 'len'] = self.end - freq.index[-1]
-
-        freq = freq.fillna(method='ffill')
-        freq['effective_rate'] = np.where(freq['state'] == 0, 0,
-                                          np.where(freq['state'] == 1, freq['rate'], float('nan')))
-        return freq
-
 
 class TraceEventCheckerBase(abc.ABC, Loggable):
     """
